requirements.txt
matplotlib
pytesseract
numpy
opencv
jupyter
pillow